+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 - Support for ESC environments in the Stack resource [#924](https://github.com/pulumi/pulumi-kubernetes-operator/pull/924)
 - Add an updateTemplate to the Stack spec [#925](https://github.com/pulumi/pulumi-kubernetes-operator/pull/925)
 - Update to Go 1.24 [#926](https://github.com/pulumi/pulumi-kubernetes-operator/pull/926)
+- More cloud logging options [#927](https://github.com/pulumi/pulumi-kubernetes-operator/pull/927)
 
 ## 2.0.0 (2025-02-18)